Enforce unique player names and birthdates
Updated the `OnModelCreating` method to add a unique index on the `Name` and `BirthDate` properties of the `Player` entity. Changed the `Name` column type in the `Players` table from `nvarchar(max)` to `nvarchar(450)` to support this index. Created a new migration class `AddUniquePlayerFeature` to apply these changes, along with updates to the corresponding designer file.